SV-9970 Add Support for Azure AD Multi-Tenant as an Identity Provider
Added code to account for Azure Multi-Tenant configuration.
Code is from an un-merged PR on the main project GitHub repo: https://github.com/openid/AppAuth-iOS/pull/714/files
The project owner seems unwilling to merge this change on the basis that a work-around is perceived to exist, and not because of any stated objection to the code. In my eyes, adding this code appears to be the best solution.